requirements.txt
dash==2.3.1
python-dotenv==0.20.0
requests==2.27.1
requests-cache==0.4.13
pandas==1.4.2
cryptography==37.0.2